Add tests and Cargo fmt
