+++ vortex-file/src/v2/exec.rs
@@ -4,7 +4,13 @@ use futures_util::StreamExt;
 use vortex_array::ArrayData;
 use vortex_error::VortexResult;
 
-use crate::v2::driver::ExecDriver;
+/// An execution driver is used to drive the execution of the scan operation.
+pub trait ExecDriver {
+    fn drive(
+        &self,
+        stream: BoxStream<'static, BoxFuture<'static, VortexResult<ArrayData>>>,
+    ) -> BoxStream<'static, VortexResult<ArrayData>>;
+}
 
 /// An execution driver that awaits the futures inline using the caller's runtime.
 pub struct InlineDriver;
