+++ vortex-file/src/v2/footer/file_layout.rs
@@ -7,7 +7,7 @@ use crate::v2::footer::segment::Segment;
 
 /// Captures the layout information of a Vortex file.
 #[derive(Clone)]
-pub(crate) struct FileLayout {
+pub struct FileLayout {
     pub(crate) root_layout: LayoutData,
     pub(crate) segments: Arc<[Segment]>,
 }
